+++ cloud/gateway-server/src/main/java/pers/acp/springcloud/gateway/controller/CustomerHystrixErrorController.java
@@ -3,9 +3,10 @@ package pers.acp.springcloud.gateway.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pers.acp.springcloud.gateway.vo.ErrorVO;
+import reactor.core.publisher.Mono;
 
 /**
  * @author zhangbin by 26/04/2018 21:30
@@ -19,12 +20,12 @@ public class CustomerHystrixErrorController {
      *
      * @return ResponseEntity
      */
-    @GetMapping(value = "/hystrixhandle", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<Object> hystrixHandle() {
+    @RequestMapping(value = "/hystrixhandle", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public Mono<ResponseEntity<Object>> hystrixHandle() {
         ErrorVO errorVO = new ErrorVO();
         errorVO.setError("invalid service");
         errorVO.setErrorDescription("GateWay error, the service is invalid");
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorVO);
+        return Mono.just(ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorVO));
     }
 
 }